Split edr functionality
